refactor: extract Linkable, Permissible, Describable concerns from Creative model
- Creative model reduced from 411 to 158 lines (-253 lines)
- Creative::Linkable: origin/linked creative logic (effective_attribute, effective_origin, etc.)
- Creative::Permissible: permission checking, children_with_permission, shared users
- Creative::Describable: description sanitization, attachment purging, effective_description
- All 13 model tests pass
